+++ lib/services/utils/logger/logger_util.dart
@@ -2,7 +2,11 @@ import 'package:logger/logger.dart';
 
 class LoggerUtil {
   static final Logger logger = Logger(
-    level: Level.info,
+    level: Level.nothing,
     printer: PrettyPrinter(),
   );
+
+  static void setLogLevel(Level level) {
+    Logger.level = level;
+  }
 }