+++ lib/services/utils/platform/i_platform_utils.dart
@@ -1,3 +1,5 @@
+import 'package:flutter/material.dart';
+
 enum PlatformType {
   mobile,
   desktop,
@@ -6,6 +8,12 @@ enum PlatformType {
 
 abstract class IPlatformUtils {
   PlatformType getPlatformType();
+
+  bool canDetectInstalledApps();
+
+  bool isBottomSheet();
+
+  bool isLongBottomSheet(Orientation orientation);
 
   bool isMobileWidth(double width);
 }