+++ lib/widgets/qr_code_widget.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:qr_flutter/qr_flutter.dart';
+import 'package:walletconnect_modal_flutter/services/utils/platform/platform_utils_singleton.dart';
 import 'package:walletconnect_modal_flutter/widgets/walletconnect_modal_theme.dart';
 
 class QRCodePage extends StatelessWidget {
@@ -14,6 +15,12 @@ class QRCodePage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    bool isLongBottomSheet = platformUtils.instance.isLongBottomSheet(
+      MediaQuery.of(context).orientation,
+    );
+    double qrSize = MediaQuery.of(context).size.height - 200;
+    double marginAndPadding = isLongBottomSheet ? 1.0 : 8.0;
+
     return Container(
       decoration: BoxDecoration(
         color: Colors.white,
@@ -21,6 +28,14 @@ class QRCodePage extends StatelessWidget {
           WalletConnectModalTheme.getData(context).radiusXS,
         ),
       ),
+      constraints: isLongBottomSheet
+          ? BoxConstraints(
+              maxHeight: qrSize,
+              maxWidth: qrSize,
+            )
+          : null,
+      margin: EdgeInsets.all(marginAndPadding),
+      padding: EdgeInsets.all(marginAndPadding),
       child: Center(
         child: QrImageView(
           data: qrData,